Improve largest scrollable element detection
In commit f405054c a special case was added for the root element:

> Consider the entire page as largest if scrollable
>
> Its area may technically be less than other elements, but if it is scrollable
> that's what you expect to scroll.

However, if you embed such a page in a frame, the page won’t be scrollable
(because it is not the uppermost root element).

This made me realize that it makes the most sense to consider the _uppermost
scrollable element_ the largest. In other words, if a scrollable element
contains another scrollable element (or a frame containing one), the parent
should be considered largest even if the child has greater area.