Fix some links not being targeted by `F`, `gf` and `gF`
In commit af0c4a07, links with an 'onclick' attribute stopped being considered
as "proper". The reason was to prevent such links from getting the same hint as
other links with the same href, because the 'onclick' likely makes all those
links behave differently.

However, on startpage.com the search result links actually are "proper" links,
but also have 'onclick' attributes and since that caused them not to be
considered as proper, they did not get hints when using `F`, `gf` and `gF`
commands.

This commit considers links with 'onclick' as proper again, and only excludes
them from getting the same hint as other links.