Fix scrolling of `<html>` if `<body>` is scrollable
`document.activeElement` is never `null`. If no element if focused, it is
`document.body`. If `document.body` is scrollable, it means that `<html>` can
never be scrolled.

This commit only allows to scroll `<body>` if it has been explicitly focused (or
if it is the largest scrollable element, of course).

Fixes #633.