Add command to click browser elements using markers
Fixes #236.

A note about the "hamburger menu" in the top-right corner:

It is possible to _open_ it with `zF`, but not possible to use `zF` inside of
it, because of `@popupPassthrough`. One _could_ an exception for that
menu/popup, but that doesn't help:

- `.elementFromPoint()` does not appear to find elements in popups, making it
  difficult to determine if the buttons inside are visible or not.
- The markers appear _behind_ the popup, no matter what `z-index` is used.
- The popup seems to eat `<escape>`: When it is pressed, the popup is closed,
  but the markers remain. The next `<escape>` press removes the markers. Not a
  big deal, but slightly annoying.

Therefore I decided not to add support for that menu in this commit. (Perhaps
we'll be able to do that some time in the future.) Instead, I recommend using
the alt key shortcuts for the "regular" menubar.