Remove try-catch in UI 'keydown' listener
Firefox unfortunately doesn't show the stack trace in its error consoles. Still,
there is one available on thrown `error` objects. That's why the try-catch used
to be there: To also show the stack trace. Wrapping the 'keydown' listener was a
good choice because most things happen on keydown in VimFx. However, since the
switch to multi-process Firefox, this is less useful. Most keys are captured in
frame script 'keydown' handlers, not in the UI, so the try-catch didn't do much.
One _could_ place the try-catch somewhere else instead, but now it's harder to
find a good place. Besides, I've not really missed the stack trace for the last
couple of months. The file name and line number is still reported which is good
enough.