Lock `:hover` when clicking with the `f` commands
Fixes #388.

Note that this commit also adds `pointer-evets: none;` to the hints container,
allowing to generate hints for elements only visible because of the mouse
pointer hovering something. Getting hints for hover-menus opened by `zf` but not
for hover-menus opened by moving the mouse was too confusing. The reason
`pointer-events: none;` wasn’t added before was that it prevented from scrolling
while hints were shown. While that was nice, it was just an edge case that
doesn't really matter. It is better to allow hints for hover-menus.