Fix misplaced hint markers when zoomed
Hint markers are supposed to be centered vertically on the non-covered point.
This didn't work properly when zoomed, but now does. For simplicity, that fix
removed the "make sure the marker stays within its element" constraint, which
turned out to be nice and more consistent. The hint markers are now _always_
centered vertically on the non-covered point, even for very small links.

This commit also fixes hint marker placement in the devtools when the devtools
have been independently zoomed. The above issue needed to be fixed before this
could be fixed. Fixes #667.