Fix a `z-index` bug for markers with the same hint
Regression since commit 1176cde8.

- `element.style.zIndex`: Always string.
- `element.style.zIndex = value`: Turns `value` into a string.
- `a = '5'; a++`: `a` is now `6`.
- `element.style.zIndex++`: If the previous value was `'5'`, the new value is
  `'6'`. This method was used before above commit.
- `element.style.zIndex += 1`: If the previous value was `'5'`, the new value is
  `'51'`. This was introduced by above commit.

The above really shows that you should always be explicit about types. Since
`element.style.zIndex` is always a string, `element.style.zIndex++` was a
terrible hack that happened to work, but caused a bug when it was rewritten to
a seemingly equivalent expression.

This commit first reads the `z-index` of the element and explicitly makes it a
`Number`, allowing incrementing to work as expected.