Fix `@popupPassThrough` for menus in frames
When a menu, such as a context menu is open, VimFx is automatically disabled to
allow using the access keys of the open menu, as well as pressing `<escape>` to
close the menu. However, this used not to work for the context menus in the
devtools and in about:config. Now it does.

This also fixes a bug where VimFx stopped being automatically disabled after
closing a sub-menu of a context menu.