Fix `zF` double clicking on browser elements
nsIDOMWindowUtils.dispatchDOMEventViaPresShell() dispatches
a 'command' event automatically after a 'click' event on XUL pages.
This was causing zF to effectively double click on browser elements,
since VimFx also simulate a 'command' event.

Elements which suffers from this double click include but are not
limited to, the new tab button, question mark circles in the preference
page and buttons created by the add-on sdk.