Teach the `zF` command to find and activate dropmarkers
For example, the bookmarks button has a "star" button on the left, which
acts like a normal button, and another button on the right, which
actually is a `<dropmarker>` and opens a `<menupopup>`. Previously,
VimFx didn't find the `<dropmarker>` (because it hasn't got
`.tabIndex > `-1). Now it does. Note that simulating a click on a
`<dropmarker>` triggers its parent button's action. Instead, the
`<menupopup>` must be found and triggered manually.

Another example of such a button is Greasemonkey’s button. Actually,
this "button with dropmarker" concept seems pretty common among add-ons.

This commit also removes the unnecessary double hint for the bookmarks
button, but filtering out `<toolbarbutton>`s which are direct children
of another `<toolbarbutton>`.

Fixes #766.