Fix unmousable bottom-right corner of the page scrollbars
VimFx's statuspanel (for notifications) had `pointer-events: auto;`
instead of `pointer-events: none;`, which statuspanels have by default.
The reason was to allow clicking the statuspanel to hide it.

This made it impossible to interact with the bottom-right corner of the
page scrollbars. The invisible statuspanel captured any clicks there.

This commit hides notifications when clicking _anything_ instead, and
removes `pointer-events: auto;`.

Fixes #787.