Make hint markers smaller
The hint markers are styled with `font: menu;` to get as close to the
user's system and preferences as possible. However, that turns out to be
slighly too big on all systems I've seen.

This commit still uses `font: menu;`, but then reduces the font size of
that to 80%, by using a clever CSS trick: `font: menu;` was moved to the
hints container, and the markers themselves use `font-size: 0.8em;`.