Properly hide notifications
In order to show the characters you have typed so far when filtering
hint markers by element text, commit ae106db9 changed the logic for when
notifications are hidden due to key presses.

Previously, notifications were hidden unconditionally when the pressed
key is not part of any shortcut, or caused a command to be run. Now, it
is up to every mode to deal with those cases. So that logic has
essentially been moved into Normal mode, and partly into Marks mode and
Caret mode as well.

The `<escape>` command now also explicitly hides notifications.