Improve hints after filtering by one element text char
1. Go to Reddit or Hackernews.
2. Press `f`.
3. Filter by one common element text character, such as "e".

Previously, that resulted in really terrible hints for most (of the
larger) elements. Basically, the difference in weight used for the
elements was so large, that all hint chars except the last were used as
single-char hints for the highest-weight elements. The same thing
happened for the hints starting with the last hint char; all hint chars
except the last were used for two-char hints. This resulted in some
really terrible hints like "VVVS". This looked very broken.

With this commit, the smallest elements still get the best hints after
filtering by element text, but the difference in the hints is not that
large.