Simplify styling for highlighted markers
Let's try _only_ changing the background color, to bright green. This
particular green does stand out compared to other markers, and does not
obscure the text, in my opinion. Green is also commonly associated with
success -- in this case, having successfully filtered your way to a
specific link.

To make the above easier, I replaced the border and matched-char colors
with semi-transparent black, so that they work on the green background
as well.